Added auth to blog comments & Gallery

+++ routes/gallery.js
@@ -15,31 +15,39 @@ router.get('/', async (req, res) => {
 // ----------- UPLOAD NEW IMAGE ---------
 
 router.get('/new', (req, res) => {
-    res.render("./../views/gallery/new", {loggedIn: req.isAuthenticated()} )
+    res.render("./../views/gallery/new", { loggedIn: req.isAuthenticated() } )
 })
 
 router.post('/new', upload.single('image'), (req, res) => {
-    let filename = ''
-    if(req.file && req.file.filename !== '') {
-        filename = req.file.filename
+    if(req.isAuthenticated()) {
+        let filename = ''
+        if(req.file && req.file.filename !== '') {
+            filename = req.file.filename
+        } else {
+            res.redirect("/gallery/new")
+        }
+    
+        const newImage = new Image({
+            image: filename
+        })
+    
+        newImage.save()
+        res.redirect("/gallery")
     } else {
-        res.redirect("/gallery/new")
+        res.redirect('/login')
     }
-
-    const newImage = new Image({
-        image: filename
-    })
-
-    newImage.save()
-    res.redirect("/gallery")
 })
 
 function deleteImage(imgUrl) {
-    fs.unlink(path.join("uploads/", imgUrl), (err) => {
-        if(err) {
-            throw err
-        }
-    })
+    if(req.isAuthenticated()) {
+        fs.unlink(path.join("uploads/", imgUrl), (err) => {
+            if(err) {
+                throw err
+            }
+        })
+    } else {
+        res.redirect("/gallery")
+    }
 }
 
 module.exports = router;
